requirements.txt
Flask==2.0.2
Werkzeug==2.0.3
python-docx==0.8.11
flask-cors==3.0.10